requirements.txt
flask
requests
pandas
python-telegram-bot==13.15